Fix crash when guessing function return types.
esdoc try to guess function return type when not `@return` tag found. When return type is Object, most times it can determine it's shape from AST. But when this object contains rest spread operator, it became compicated (for ex `return { ...obj }`, theorethically we can try track deeper and determine shape of `obj`, but for now we simple break guessing with `*`). In any case, this commit fixed crash with this case, because when AST node contain object with type `SpreadProperty`, it of course has no `key`.